Revert "Gerrit queries don't always seem to work"
This reverts commit cbece4b52cf006ddb189e8bfd54f44c5f145ae86.

This change completely broke querying. For example, for the
invocation

 qgerrit --status abandoned --status merged \
         --user vishvananda --user berrange \
         --project openstack/nova

which says "given me all patches by vish or berrange,
and which are merged or abandoned, and are in the nova
project. This generates a query string of

   (owner:berrange AND (project:openstack/nova OR
    status:abandoned OR status:merged)) OR
   (owner:vishvananda AND (project:openstack/nova OR
    status:abandoned OR status:merged)) OR
   (project:openstack/nova AND (owner:berrange OR
    owner:vishvananda OR status:abandoned OR status:merged))
    OR (status:abandoned AND (owner:berrange OR
    owner:vishvananda OR project:openstack/nova)) OR
   (status:merged AND (owner:berrange OR owner:vishvananda
   OR project:openstack/nova))

which is totally mangled and returns incorrect results,
for examples, shows reviewers by other users

Reverting it puts the query back to

  (owner:vishvananda OR owner:berrange) AND \
  (status:abandoned OR status:merged) AND \
  (project:openstack/nova)

ie, values for each type of arg are taken as a union, while it takes
the intersection of each different type, as per the description
of desired results.