allow ssh private key to come from local agent by default
This patch defaults opts.keyfile to None.  This will cause the paramiko
library to attempt to use a local ssh-agent, which matches the default
behavior of ssh (and hence git, et al).

This is important because there is no guarantee that the system on which
someone is running cgerrit has access to the actual private key file.