More improvements to filtering on approval score
The previous logic for filtering approval score was not working
well because it required that all approval values were matching.

eg  c2, required that every single code review was +2.

What is more useful is to say that at least one code review
must be +2, and all others greater than 0.

This change makes the logic be:

  # approval == {v,c,a}NN  (v == verified, c == code review, a == approved)
  #
  # NN == -1 -> patches without any -2
  # NN == 0 -> patches without any -1 or -2
  # NN == 1 -> patches without any -1 or -2 and at least one +1
  # NN == 2 -> patches without any -1 or -2 and at least one +2

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>